requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
filelock==3.19.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2025.9.0
gunicorn==23.0.0
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
PyPDF2==3.0.1
pytest==8.4.2
pytest-flask==1.3.0
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
SQLAlchemy==2.0.43
sympy==1.14.0
tokenizers==0.22.1
tomli==2.2.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.23.0
